Update user 3 (#5)
* добавить класс UpdateUserRequest;

* изменить текс логирования в UserController;

* добавить метод updateUser в UserMapper;

* добавить метод updateUser;

* добавить класс UserShortDto;

* добавить класс UserAdminController и перенести в него часть методов;

* добавить метод deleteUser в UserAdminController;

* добавить метод getUserById в UserAdminController;

* изменить комментарии ToDo и текст в логировании;

* добавить метод userToUserShortDto в UserMapper;

* добавить метод userListToUserShortDtoList в UserMapper;

* добавить enum Role;

* добавить поле role и исправить поле network на Set<String> в User;

* добавить класс BadRequestException;

* исправить метод deleteUser в UserController;

* исправить метод getUserById в UserController;

* исправить метод updateUser в UserController;

* добавить класс ErrorHandler;

+++ src/main/java/com/kardoaward/kardo/user/controller/UserController.java
@@ -1,12 +1,13 @@
 package com.kardoaward.kardo.user.controller;
 
+import com.kardoaward.kardo.exception.BadRequestException;
 import com.kardoaward.kardo.user.mapper.UserMapper;
 import com.kardoaward.kardo.user.model.dto.NewUserRequest;
 import com.kardoaward.kardo.user.model.User;
+import com.kardoaward.kardo.user.model.dto.UpdateUserRequest;
 import com.kardoaward.kardo.user.model.dto.UserDto;
 import com.kardoaward.kardo.user.service.UserService;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,15 +15,14 @@ import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @Slf4j
 @AllArgsConstructor
@@ -36,41 +36,70 @@ public class UserController {
     private final UserMapper userMapper;
 
     @PostMapping
-    //ToDo Какой статус возвращать фронту и нужно ли?
+    /*ToDo
+       Какой статус возвращать фронту и нужно ли?
+     */
     @ResponseStatus(HttpStatus.CREATED)
     public UserDto createUser(@RequestBody @Valid NewUserRequest newUserRequest) {
-        log.info("Получен новый пользователь {}.", newUserRequest);
+        log.info("Добавление нового пользователь {}.", newUserRequest);
         User user = userMapper.newUserRequestToUser(newUserRequest);
         User returnedUser = userService.addUser(user);
         return userMapper.userToUserDto(returnedUser);
     }
 
     @GetMapping("/{userId}")
-    //ToDo Какой статус возвращать фронту и нужно ли?
+    /* ToDo
+        Какой статус возвращать фронту и нужно ли?
+     */
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public UserDto getUserById(@PathVariable @Positive Long userId) {
-        //ToDo Изменить текст лога?
-        log.info("Запрос на возврат пользователя с ИД {}.", userId);
+    public UserDto getUserById(@RequestHeader("X-Participant-User-Id") Long participantId,
+                               @PathVariable @Positive Long userId) {
+        log.info("Получение пользователем с ИД {} своих данных.", userId);
+
+        if (!userId.equals(participantId)) {
+            log.error("Пользователь с ИД {} не может просматривать профиль пользователя с ИД {}.", participantId, userId);
+            throw new BadRequestException(String.format("Пользователь с ИД %d не может просматривать " +
+                    "профиль пользователя с ИД %d.", participantId, userId));
+        }
+
         User returnedUser = userService.getUserById(userId);
         return userMapper.userToUserDto(returnedUser);
     }
 
     @DeleteMapping("/{userId}")
-    //ToDo Какой статус возвращать фронту и нужно ли?
+    /* ToDo
+        Какой статус возвращать фронту и нужно ли?
+     */
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable @Positive Long userId) {
-        //ToDo Изменить текст лога?
-        log.info("Запрос на удаление пользователя с ИД {}.", userId);
+    public void deleteUser(@RequestHeader("X-Participant-User-Id") Long participantId,
+                           @PathVariable @Positive Long userId) {
+        log.info("Удаление пользователем с ИД {} своего профиля.", userId);
+
+        if (!userId.equals(participantId)) {
+            log.error("Пользователь с ИД {} не может удалить профиль пользователя с ИД {}.", participantId, userId);
+            throw new BadRequestException(String.format("Пользователь с ИД %d не может удалить " +
+                    "профиль пользователя с ИД %d.", participantId, userId));
+        }
+
         userService.deleteUser(userId);
     }
 
-    @GetMapping
-    //ToDo Нужен ли такой метод? Если нет, то переделать под получение всех пользователей
-    public List<UserDto> getUsersByIds(@RequestParam(required = false) List<Long> ids,
-                                       @RequestParam(defaultValue = "0") @Min(0) int from,
-                                       @RequestParam(defaultValue = "10") @Positive int size) {
-        log.info("Запрос на возврат списка пользователей.");
-        List<User> users = userService.getUsersByIds(ids, from, size);
-        return userMapper.userListToUserDtoList(users);
+    @PatchMapping("/{userId}")
+    /* ToDo
+        Какой статус возвращать фронту и нужно ли?
+     */
+    public UserDto updateUser(@RequestHeader("X-Participant-User-Id") Long participantId,
+                              @PathVariable @Positive Long userId,
+                              @RequestBody @Valid UpdateUserRequest request) {
+        log.info("Обновление пользователем с ИД {} своих данных.", userId);
+
+        if (!userId.equals(participantId)) {
+            log.error("Пользователь с ИД {} не может обновить профиль пользователя с ИД {}.", participantId, userId);
+            throw new BadRequestException(String.format("Пользователь с ИД %d не может обновить " +
+                    "профиль пользователя с ИД %d.", participantId, userId));
+        }
+
+        User updatedUser = userService.updateUser(userId, request);
+        return userMapper.userToUserDto(updatedUser);
     }
 }
